+++ internal/provider/dashboards/dashboard_schema/v1.go
@@ -1574,7 +1574,6 @@ func dashboardSchemaAttributesV1() map[string]schema.Attribute {
 			Attributes: map[string]schema.Attribute{
 				"id": schema.StringAttribute{
 					Optional: true,
-					Computed: true,
 					Validators: []validator.String{
 						stringvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("path"),
@@ -1583,7 +1582,6 @@ func dashboardSchemaAttributesV1() map[string]schema.Attribute {
 				},
 				"path": schema.StringAttribute{
 					Optional: true,
-					Computed: true,
 					Validators: []validator.String{
 						stringvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("id"),