+++ internal/provider/dashboards/dashboard_schema/v3.go
@@ -1420,7 +1420,6 @@ func dashboardSchemaAttributesV3() map[string]schema.Attribute {
 			Attributes: map[string]schema.Attribute{
 				"id": schema.StringAttribute{
 					Optional: true,
-					Computed: true,
 					Validators: []validator.String{
 						stringvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("path"),
@@ -1429,7 +1428,6 @@ func dashboardSchemaAttributesV3() map[string]schema.Attribute {
 				},
 				"path": schema.StringAttribute{
 					Optional: true,
-					Computed: true,
 					Validators: []validator.String{
 						stringvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("id"),